server/requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==10.3.0
pycparser==2.22
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypdfium2==4.30.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.1
Werkzeug==3.0.3
wsproto==1.2.0
